Rewrite the unpacker, the packer and the server beneath their pins
The last three files of the batch. `unpacker/mod.rs`'s four exports
(`unpack`, `unpacker_{init,advance,teardown}`) keep their signatures and
their `#[unsafe(no_mangle)]` exactly; everything moved is beneath them.
`extern_abi` 759, `no_mangle` 322 and the ABI ledger's 322 entries are
flat.

Four levers, in the order they paid:

1. **`Cursor`'s construction is the unsafe step.** `Cursor::new(data,
   size)` is the promise that the bytes are readable; `next()` and
   `take()` are safe after it, and with them `parse_redraw` and
   `parse_grid_line_cell` become **safe fns** — the whole redraw state
   machine, ~250 lines, costs 32 unchecked lines now.
2. **Destructuring `&mut Unpacker`.** `let Unpacker { grid_line_event:
   g, state, nevents, .. } = u;` splits one borrow into disjoint field
   borrows, which is what let the redraw decoder stop carrying a raw
   `*mut GridLineEvent` beside the unpacker it points into.
3. **`scalar_object(tok) -> Option<Object>`**, a *pure* fn: a token is a
   value and *building* a union is safe, only reading one back is not.
   That is four arms of `api_parse_enter` off the unchecked count, and
   it is directly testable.
4. **A `log!` macro in `msgpack_rpc/mod.rs`.** Spelled out, `logmsg_c!`
   is six fixed arguments before the format string and wraps over eight
   lines — eight lines of unchecked code, since the whole call sits
   inside the region. Naming the fixed half leaves one line per site.
   `server.rs` has seven of them and fell 241 -> 166 on this alone.

`unpacker_advance` deliberately keeps raw-pointer discipline: it calls
`mpack_parse`, which re-enters the unpacker through `parser.data.p`, so
a `&mut Unpacker` held across it would strip the callback's permission
to write what it then reads. `unpacker_parse_header` does take one —
nothing it calls comes back.

Three files became four plus two, because SAFETY notes took both over
the 1,000-line cap:

- `unpacker/keydict.rs` — `unpack_keydict` and the cursor primitives
  (`unpack_{string,array,integer,skip}`, `push_additional_data`). Worth
  saying plainly: **none of this is on the RPC path.** Its only caller
  is ShaDa, and `perssweep` is what watches it.
- `unpacker/redraw.rs` — the `grid_line` fast path and the `Cursor`.

Both are new files the ratchet scores from zero, hence `--allow-growth`;
with `channel/receive.rs` that is three, and they raise
`files_without_forbid_unsafe` by three for the same reason.

**All twelve files carry both markers now**, the cast deny included, and
`msgpack_rpc/mod.rs` can carry it because the subtree beneath it is
clean. Two casts needed a decision rather than a spelling:

- `(uint32_t)mpack_unpack_uint(tok)` for the envelope's message type and
  request id is load-bearing wrapping — a message type of
  `0x1_0000_0000` decodes as 0, i.e. a *request* — so it is
  `crate::narrow::msgpack_uint_as_u32` now, beside `number_as_int`, with
  a Miri test.
- `packer/format.rs`'s twenty width narrowings are all guarded by the
  `if` immediately above them (that is what choosing a msgpack width
  *means*), so they are `try_from(..).expect(..)` with one message
  naming the guard. `MPACK_{OK,EOF,ERROR}` are `c_int` rather than
  `c_uint`, which deletes eleven casts and two in `shada/`;
  `unit-cdefs.h` already rendered them as `static const int`.

Six new `#[cfg(test)]` tests over the `Cursor` and `scalar_object`,
green under Miri — the token reader's pointer arithmetic is exactly what
Miri is for.

Batch unchecked 1,872 -> 1,195 (-36 %): unpacker 711 -> 233 + 191 + 32,
packer 165 -> 62, server 241 -> 166, channel 755 -> 511. Files carrying
neither safety marker 41 -> 34 (the batch is closed); files without the
cast deny 1,095 -> 1,088 with three new files in the denominator;
`missing_safety_doc` 3,973 -> 3,878. clippy stays at 20/17.

decodediff all five sections IDENTICAL against `c6eea31263`.

+++ crates/nvim/src/shada/parse.rs
@@ -9,7 +9,7 @@
 #![deny(unsafe_op_in_unsafe_fn)]
 
 use crate::semsg_c;
-use core::ffi::{CStr, c_char, c_int, c_uint, c_void};
+use core::ffi::{CStr, c_char, c_int, c_void};
 
 use crate::msgpack_rpc::unpacker::MPACK_OK;
 
@@ -483,7 +483,7 @@ unsafe fn parse_variable(
                 is_blob = true;
             }
             (*global_var).value = decode_string(binval.data(), binval.len(), is_blob, false);
-        } else if cursor.typval(&raw mut (*global_var).value) as c_uint != MPACK_OK {
+        } else if cursor.typval(&raw mut (*global_var).value) != MPACK_OK {
             semsg_c!(
                 gettext(
                     c"E575: Error while reading ShaDa file: variable entry at position %lu has value that cannot be converted to the Vimscript value"
